+++ tests/Jieba/MultiArrayTest.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Jieba\Tests\Jieba;
+
 use Jieba\MultiArray;
 use PHPUnit\Framework\TestCase;
 
@@ -72,5 +74,94 @@ class MultiArrayTest extends TestCase
     public function testExists(bool $expected, array $array, string $keyDelimiter, string $keyString, string $message)
     {
         $this->assertSame($expected, (new MultiArray($array, $keyDelimiter))->exists($keyString), $message);
+    }
+
+    /**
+     * @return array
+     */
+    public function dataRemove(): array
+    {
+        $array = [
+            'key11' => [
+                'key21' => '',
+                'key22' => null,
+                'key23' => [
+                    'key31' => true,
+                    'key32' => [
+                        'key41' => false,
+                        'key42' => true,
+                    ],
+                ],
+            ],
+        ];
+
+        return [
+            [
+                $array,
+                'key11.key21',
+                [
+                    'key11.key21',
+                ],
+                [
+                    'key11.key23',
+                    'key11.key23.key31',
+                    'key11.key23.key32',
+                    'key11.key23.key32.key41',
+                ],
+            ],
+            [
+                $array,
+                'key11.key23',
+                [
+                    'key11.key23',
+                ],
+                [
+                    'key11.key21',
+
+                    // NOTE: I thought following keys should be removed as well but not.
+                    // TODO: why following keys not removed after \Jieba\MultiArray::remove() has been called?
+                    'key11.key23.key31',
+                    'key11.key23.key32',
+                    'key11.key23.key32.key41',
+                ],
+            ],
+        ];
+    }
+
+    /**
+     * @dataProvider dataRemove
+     * @depends testExists
+     * @covers \Jieba\MultiArray::exists()
+     * @covers \Jieba\MultiArray::remove()
+     * @covers \Jieba\MultiArray::set()
+     * @param array $array
+     * @param string $deletedKeyString
+     * @param array $deletedKeyStrings
+     * @param array $existingKeyStrings
+     * @return void
+     */
+    public function testRemove(
+        array $array,
+        string $deletedKeyString,
+        array $deletedKeyStrings,
+        array $existingKeyStrings
+    ) {
+        $a = new MultiArray($array);
+
+        foreach ($existingKeyStrings as $keyString) {
+            $this->assertTrue($a->exists($keyString), "key {$keyString} should be in the array before removed");
+        }
+        foreach ($deletedKeyStrings as $keyString) {
+            $this->assertTrue($a->exists($keyString), "key {$keyString} should be in the array before removed");
+        }
+
+        $a->remove($deletedKeyString);
+
+        foreach ($existingKeyStrings as $keyString) {
+            $this->assertTrue($a->exists($keyString), "key {$keyString} should still be there");
+        }
+        foreach ($deletedKeyStrings as $keyString) {
+            $this->assertFalse($a->exists($keyString), "key {$keyString} should have been removed from the array");
+        }
     }
 }